added fixes to issuers moved the user context for each database
