+++ go.mod
@@ -5,7 +5,7 @@ go 1.19
 require (
 	github.com/cjlapao/common-go v0.0.35
 	github.com/cjlapao/common-go-cryptorand v0.0.4
-	github.com/cjlapao/common-go-database v0.0.3
+	github.com/cjlapao/common-go-database v0.0.4
 	github.com/cjlapao/common-go-restapi v0.0.6
 	github.com/google/uuid v1.3.0
 	github.com/gorilla/mux v1.8.0
